requirements.txt
accelerate @ file:///home/conda/feedstock_root/build_artifacts/accelerate_1734379448962/work
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///croot/async-timeout_1732661977001/work
attrs @ file:///croot/attrs_1734533101012/work
bitsandbytes==0.45.0
bitsandbytes-cuda117==0.26.0.post2
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1734473278428/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
colorama @ file:///croot/colorama_1672386526460/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1733931864043/work
dill @ file:///croot/dill_1715094664823/work
docstring_parser @ file:///home/conda/feedstock_root/build_artifacts/docstring_parser_1710507861579/work
eval_type_backport @ file:///home/conda/feedstock_root/build_artifacts/eval_type_backport_1734857727836/work
filelock @ file:///croot/filelock_1700591183607/work
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec @ file:///croot/fsspec_1714461537038/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1733852339297/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1730902924303/work
joblib @ file:///croot/joblib_1718217211762/work
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mdurl @ file:///opt/conda/conda-bld/mdurl_1659716024347/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
networkx @ file:///croot/networkx_1717597493534/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp39-cp39-linux_x86_64.whl#sha256=0a51ed2e81ab863e3d00ed6c5049192ce578ecb38fb467d2f9a6585d3c25f666
peft @ file:///home/conda/feedstock_root/build_artifacts/peft_1733652333873/work
pillow @ file:///croot/pillow_1734430590456/work
propcache @ file:///croot/propcache_1732303986938/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pyarrow @ file:///croot/pyarrow_1733934515071/work/python
Pygments @ file:///croot/pygments_1684279966437/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv @ file:///croot/python-dotenv_1669132557191/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1726669635272/work
requests @ file:///croot/requests_1730999120400/work
rich @ file:///croot/rich_1732638981168/work
safetensors @ file:///croot/safetensors_1732227615904/work
scikit-learn @ file:///croot/scikit-learn_1734547417863/work
scipy @ file:///croot/scipy_1733756309941/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=a4372e07d982a377bda57ec71fd74ac403442571e1d717de47325a14dea5c278
shtab @ file:///home/conda/feedstock_root/build_artifacts/shtab_1734664526617/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///croot/sympy_1734622612703/work
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tokenizers @ file:///home/conda/feedstock_root/build_artifacts/tokenizers_1732733942280/work/bindings/python
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm @ file:///croot/tqdm_1724853939799/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1734498907751/work
triton==2.0.0
trl @ file:///home/conda/feedstock_root/build_artifacts/trl_1734311638671/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tyro @ file:///home/conda/feedstock_root/build_artifacts/tyro_1732010820374/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1727769808118/work
xxhash @ file:///croot/python-xxhash_1667919521615/work
yarl @ file:///croot/yarl_1732546845924/work

pandas~=2.2.3
numpy~=1.26.4
transformers~=4.47.1
python-dotenv~=0.21.0
scikit-learn~=1.5.2
tqdm~=4.66.5
peft~=0.14.0
trl~=0.13.0
datasets~=3.2.0
wandb~=0.19.4